requirements.txt
fastapi~=0.110.1
uvicorn
sqlalchemy~=2.0.29
psycopg2-binary
passlib~=1.7.4
python-jose